validateCGC and parseCGC working
